lab53/requirements.txt
asgiref==3.4.1
astroid==2.8.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
configparser==5.0.2
Django==3.2.8
django-crispy-forms==1.13.0
dnspython==2.1.0
greenlet==1.1.2
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
mongoengine==0.23.1
mysql-connector-python==8.0.26
pass-admin-pkg-Smarv==0.0.1
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.17.3
pylint==2.11.1
pymongo==3.12.0
PyMySQL==1.0.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.19.0
pytz==2021.3
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
websockets==10.0
wrapt==1.12.1
